+++ marusia/errors.go
@@ -1,8 +1,0 @@
-package marusia
-
-import "errors"
-
-var (
-	TooLongRunning    = errors.New("too long run scene")
-	BadDirectorAnswer = errors.New("bad answer from script director")
-)